Cleaned up the STM32F4 tree and removed references to the encoder and cam make
This will allow further development of things such as large UART read/writes as
the STM32F4 board has much larger memory space (128KB) versus the 16KB of the
STM32L1 board.

The code is still there in the tree for the encoder and camera interface... they
should really be kept concurrent between the STM32F4 and STM32L1 trees.

-MD